Can't delete the second user form the app

+++ lib/modules/auth/blocs/auth_bloc/auth_methods.dart
@@ -107,12 +107,14 @@ class AuthMethods {
     return _accountsDao.getAccounts(user.localId);
   }
 
-  Future<void> logout(User user) async {
+  Future<void> logout(int currentUserId, User user) async {
     _authApi.logout(user);
 
     final futures = [
       deleteUserRelatedData(user),
+      if(user.localId==currentUserId)
       _authLocal.deleteSelectedUserLocalId(),
+      if(user.localId==currentUserId)
       _authLocal.deleteSelectedAccountId(),
       _usersDao.deleteUser(user.localId),
       _accountsDao.deleteAccountsOfUser(user.localId),
